requirements.txt
Flask>=2.0
python-dotenv>=0.20
supabase==2.20.0
